requirements.txt
python-telegram-bot==21.5
ollama==0.3.3
python-dotenv==1.0.1
Flask==3.0.3
google-api-python-client==2.170.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
